1.店长系统优化 2.手册系统开发 3.角色系统修改 4.书境系统修改

+++ ZY-APP-BASE/src/com/zywl/app/base/bean/UserRole.java
@@ -14,7 +14,7 @@ public class UserRole extends BaseBean {
 
     private int index;
 
-    private int hp;
+    private double hp;
 
     private int maxHp;
 
@@ -54,11 +54,11 @@ public class UserRole extends BaseBean {
         this.roleId = roleId;
     }
 
-    public int getHp() {
+    public double getHp() {
         return hp;
     }
 
-    public void setHp(int hp) {
+    public void setHp(double hp) {
         this.hp = hp;
     }
